12.6.6.0

+++ androidx/core/graphics/TypefaceCompatUtil.java
@@ -12,6 +12,12 @@ import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 import androidx.annotation.RestrictTo;
 import com.baidu.mapsdkplatform.comapi.map.r;
+import com.baidu.mobads.container.util.AdIconUtil;
+import com.baidu.titan.sdk.runtime.FieldHolder;
+import com.baidu.titan.sdk.runtime.InitContext;
+import com.baidu.titan.sdk.runtime.InterceptResult;
+import com.baidu.titan.sdk.runtime.Interceptable;
+import com.baidu.titan.sdk.runtime.TitanRuntime;
 import java.io.Closeable;
 import java.io.File;
 import java.io.FileInputStream;
@@ -24,47 +30,75 @@ import java.nio.channels.FileChannel;
 @RestrictTo({RestrictTo.Scope.LIBRARY_GROUP_PREFIX})
 /* loaded from: classes.dex */
 public class TypefaceCompatUtil {
+    public static /* synthetic */ Interceptable $ic = null;
     public static final String CACHE_FILE_PREFIX = ".font";
     public static final String TAG = "TypefaceCompatUtil";
+    public transient /* synthetic */ FieldHolder $fh;
+
+    public TypefaceCompatUtil() {
+        Interceptable interceptable = $ic;
+        if (interceptable != null) {
+            InitContext newInitContext = TitanRuntime.newInitContext();
+            interceptable.invokeUnInit(65536, newInitContext);
+            int i2 = newInitContext.flag;
+            if ((i2 & 1) != 0) {
+                int i3 = i2 & 2;
+                newInitContext.thisArg = this;
+                interceptable.invokeInitBody(65536, newInitContext);
+            }
+        }
+    }
 
     public static void closeQuietly(Closeable closeable) {
-        if (closeable != null) {
-            try {
-                closeable.close();
-            } catch (IOException unused) {
-            }
+        Interceptable interceptable = $ic;
+        if (!(interceptable == null || interceptable.invokeL(65537, null, closeable) == null) || closeable == null) {
+            return;
+        }
+        try {
+            closeable.close();
+        } catch (IOException unused) {
         }
     }
 
     @Nullable
     @RequiresApi(19)
     public static ByteBuffer copyToDirectBuffer(Context context, Resources resources, int i2) {
-        File tempFile = getTempFile(context);
-        if (tempFile == null) {
-            return null;
-        }
-        try {
-            if (copyToFile(tempFile, resources, i2)) {
-                return mmap(tempFile);
+        InterceptResult invokeLLI;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeLLI = interceptable.invokeLLI(65538, null, context, resources, i2)) == null) {
+            File tempFile = getTempFile(context);
+            if (tempFile == null) {
+                return null;
             }
-            return null;
-        } finally {
-            tempFile.delete();
+            try {
+                if (copyToFile(tempFile, resources, i2)) {
+                    return mmap(tempFile);
+                }
+                return null;
+            } finally {
+                tempFile.delete();
+            }
         }
+        return (ByteBuffer) invokeLLI.objValue;
     }
 
     public static boolean copyToFile(File file, InputStream inputStream) {
+        InterceptResult invokeLL;
         FileOutputStream fileOutputStream;
+        Interceptable interceptable = $ic;
+        if (interceptable != null && (invokeLL = interceptable.invokeLL(65540, null, file, inputStream)) != null) {
+            return invokeLL.booleanValue;
+        }
         StrictMode.ThreadPolicy allowThreadDiskWrites = StrictMode.allowThreadDiskWrites();
         FileOutputStream fileOutputStream2 = null;
         try {
             try {
                 fileOutputStream = new FileOutputStream(file, false);
-            } catch (IOException e2) {
-                e = e2;
+            } catch (Throwable th) {
+                th = th;
             }
-        } catch (Throwable th) {
-            th = th;
+        } catch (IOException e2) {
+            e = e2;
         }
         try {
             byte[] bArr = new byte[1024];
@@ -96,47 +130,59 @@ public class TypefaceCompatUtil {
 
     @Nullable
     public static File getTempFile(Context context) {
-        File cacheDir = context.getCacheDir();
-        if (cacheDir == null) {
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(AdIconUtil.AD_TEXT_ID, null, context)) == null) {
+            File cacheDir = context.getCacheDir();
+            if (cacheDir == null) {
+                return null;
+            }
+            String str = CACHE_FILE_PREFIX + Process.myPid() + "-" + Process.myTid() + "-";
+            for (int i2 = 0; i2 < 100; i2++) {
+                File file = new File(cacheDir, str + i2);
+                if (file.createNewFile()) {
+                    return file;
+                }
+            }
             return null;
         }
-        String str = CACHE_FILE_PREFIX + Process.myPid() + "-" + Process.myTid() + "-";
-        for (int i2 = 0; i2 < 100; i2++) {
-            File file = new File(cacheDir, str + i2);
-            if (file.createNewFile()) {
-                return file;
-            }
-        }
-        return null;
+        return (File) invokeL.objValue;
     }
 
     @Nullable
     @RequiresApi(19)
     public static ByteBuffer mmap(File file) {
-        try {
-            FileInputStream fileInputStream = new FileInputStream(file);
-            FileChannel channel = fileInputStream.getChannel();
-            MappedByteBuffer map = channel.map(FileChannel.MapMode.READ_ONLY, 0L, channel.size());
-            fileInputStream.close();
-            return map;
-        } catch (IOException unused) {
-            return null;
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(65543, null, file)) == null) {
+            try {
+                FileInputStream fileInputStream = new FileInputStream(file);
+                FileChannel channel = fileInputStream.getChannel();
+                MappedByteBuffer map = channel.map(FileChannel.MapMode.READ_ONLY, 0L, channel.size());
+                fileInputStream.close();
+                return map;
+            } catch (IOException unused) {
+                return null;
+            }
         }
+        return (ByteBuffer) invokeL.objValue;
     }
 
     @Nullable
     @RequiresApi(19)
     public static ByteBuffer mmap(Context context, CancellationSignal cancellationSignal, Uri uri) {
-        try {
-            ParcelFileDescriptor openFileDescriptor = context.getContentResolver().openFileDescriptor(uri, r.f7715a, cancellationSignal);
-            if (openFileDescriptor == null) {
-                if (openFileDescriptor != null) {
-                    openFileDescriptor.close();
-                }
-                return null;
-            }
-            FileInputStream fileInputStream = new FileInputStream(openFileDescriptor.getFileDescriptor());
+        InterceptResult invokeLLL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeLLL = interceptable.invokeLLL(AdIconUtil.BAIDU_LOGO_ID, null, context, cancellationSignal, uri)) == null) {
             try {
+                ParcelFileDescriptor openFileDescriptor = context.getContentResolver().openFileDescriptor(uri, r.f7745a, cancellationSignal);
+                if (openFileDescriptor == null) {
+                    if (openFileDescriptor != null) {
+                        openFileDescriptor.close();
+                    }
+                    return null;
+                }
+                FileInputStream fileInputStream = new FileInputStream(openFileDescriptor.getFileDescriptor());
                 FileChannel channel = fileInputStream.getChannel();
                 MappedByteBuffer map = channel.map(FileChannel.MapMode.READ_ONLY, 0L, channel.size());
                 fileInputStream.close();
@@ -144,21 +190,20 @@ public class TypefaceCompatUtil {
                     openFileDescriptor.close();
                 }
                 return map;
-            } catch (Throwable th) {
-                try {
-                    fileInputStream.close();
-                } catch (Throwable th2) {
-                    th.addSuppressed(th2);
-                }
-                throw th;
+            } catch (IOException unused) {
+                return null;
             }
-        } catch (IOException unused) {
-            return null;
         }
+        return (ByteBuffer) invokeLLL.objValue;
     }
 
     public static boolean copyToFile(File file, Resources resources, int i2) {
         InputStream inputStream;
+        InterceptResult invokeLLI;
+        Interceptable interceptable = $ic;
+        if (interceptable != null && (invokeLLI = interceptable.invokeLLI(65539, null, file, resources, i2)) != null) {
+            return invokeLLI.booleanValue;
+        }
         try {
             inputStream = resources.openRawResource(i2);
             try {
