+++ androidx/core/util/AtomicFile.java
@@ -3,6 +3,12 @@ package androidx.core.util;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import com.baidu.android.imsdk.internal.Constants;
+import com.baidu.titan.sdk.runtime.FieldHolder;
+import com.baidu.titan.sdk.runtime.InitContext;
+import com.baidu.titan.sdk.runtime.InterceptResult;
+import com.baidu.titan.sdk.runtime.Interceptable;
+import com.baidu.titan.sdk.runtime.TitanRuntime;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -10,69 +16,109 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 /* loaded from: classes.dex */
 public class AtomicFile {
+    public static /* synthetic */ Interceptable $ic;
+    public transient /* synthetic */ FieldHolder $fh;
     public final File mBackupName;
     public final File mBaseName;
 
     public AtomicFile(@NonNull File file) {
+        Interceptable interceptable = $ic;
+        if (interceptable != null) {
+            InitContext newInitContext = TitanRuntime.newInitContext();
+            newInitContext.initArgs = r2;
+            Object[] objArr = {file};
+            interceptable.invokeUnInit(65536, newInitContext);
+            int i2 = newInitContext.flag;
+            if ((i2 & 1) != 0) {
+                int i3 = i2 & 2;
+                newInitContext.thisArg = this;
+                interceptable.invokeInitBody(65536, newInitContext);
+                return;
+            }
+        }
         this.mBaseName = file;
         this.mBackupName = new File(file.getPath() + ".bak");
     }
 
     public static boolean sync(@NonNull FileOutputStream fileOutputStream) {
-        try {
-            fileOutputStream.getFD().sync();
-            return true;
-        } catch (IOException unused) {
-            return false;
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(65537, null, fileOutputStream)) == null) {
+            try {
+                fileOutputStream.getFD().sync();
+                return true;
+            } catch (IOException unused) {
+                return false;
+            }
         }
+        return invokeL.booleanValue;
     }
 
     public void delete() {
-        this.mBaseName.delete();
-        this.mBackupName.delete();
+        Interceptable interceptable = $ic;
+        if (interceptable == null || interceptable.invokeV(1048576, this) == null) {
+            this.mBaseName.delete();
+            this.mBackupName.delete();
+        }
     }
 
     public void failWrite(@Nullable FileOutputStream fileOutputStream) {
-        if (fileOutputStream != null) {
-            sync(fileOutputStream);
-            try {
-                fileOutputStream.close();
-                this.mBaseName.delete();
-                this.mBackupName.renameTo(this.mBaseName);
-            } catch (IOException e2) {
-                Log.w("AtomicFile", "failWrite: Got exception:", e2);
-            }
+        Interceptable interceptable = $ic;
+        if (!(interceptable == null || interceptable.invokeL(Constants.METHOD_GET_CONTACTER_INFO_FOR_SESSION, this, fileOutputStream) == null) || fileOutputStream == null) {
+            return;
+        }
+        sync(fileOutputStream);
+        try {
+            fileOutputStream.close();
+            this.mBaseName.delete();
+            this.mBackupName.renameTo(this.mBaseName);
+        } catch (IOException e2) {
+            Log.w("AtomicFile", "failWrite: Got exception:", e2);
         }
     }
 
     public void finishWrite(@Nullable FileOutputStream fileOutputStream) {
-        if (fileOutputStream != null) {
-            sync(fileOutputStream);
-            try {
-                fileOutputStream.close();
-                this.mBackupName.delete();
-            } catch (IOException e2) {
-                Log.w("AtomicFile", "finishWrite: Got exception:", e2);
-            }
+        Interceptable interceptable = $ic;
+        if (!(interceptable == null || interceptable.invokeL(Constants.METHOD_SEND_USER_MSG, this, fileOutputStream) == null) || fileOutputStream == null) {
+            return;
+        }
+        sync(fileOutputStream);
+        try {
+            fileOutputStream.close();
+            this.mBackupName.delete();
+        } catch (IOException e2) {
+            Log.w("AtomicFile", "finishWrite: Got exception:", e2);
         }
     }
 
     @NonNull
     public File getBaseFile() {
-        return this.mBaseName;
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        return (interceptable == null || (invokeV = interceptable.invokeV(1048579, this)) == null) ? this.mBaseName : (File) invokeV.objValue;
     }
 
     @NonNull
     public FileInputStream openRead() throws FileNotFoundException {
-        if (this.mBackupName.exists()) {
-            this.mBaseName.delete();
-            this.mBackupName.renameTo(this.mBaseName);
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeV = interceptable.invokeV(1048580, this)) == null) {
+            if (this.mBackupName.exists()) {
+                this.mBaseName.delete();
+                this.mBackupName.renameTo(this.mBaseName);
+            }
+            return new FileInputStream(this.mBaseName);
         }
-        return new FileInputStream(this.mBaseName);
+        return (FileInputStream) invokeV.objValue;
     }
 
     @NonNull
     public byte[] readFully() throws IOException {
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        if (interceptable != null && (invokeV = interceptable.invokeV(1048581, this)) != null) {
+            return (byte[]) invokeV.objValue;
+        }
         FileInputStream openRead = openRead();
         try {
             byte[] bArr = new byte[openRead.available()];
@@ -97,26 +143,31 @@ public class AtomicFile {
 
     @NonNull
     public FileOutputStream startWrite() throws IOException {
-        if (this.mBaseName.exists()) {
-            if (!this.mBackupName.exists()) {
-                if (!this.mBaseName.renameTo(this.mBackupName)) {
-                    Log.w("AtomicFile", "Couldn't rename file " + this.mBaseName + " to backup file " + this.mBackupName);
-                }
-            } else {
-                this.mBaseName.delete();
-            }
-        }
-        try {
-            return new FileOutputStream(this.mBaseName);
-        } catch (FileNotFoundException unused) {
-            if (this.mBaseName.getParentFile().mkdirs()) {
-                try {
-                    return new FileOutputStream(this.mBaseName);
-                } catch (FileNotFoundException unused2) {
-                    throw new IOException("Couldn't create " + this.mBaseName);
+        InterceptResult invokeV;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeV = interceptable.invokeV(1048582, this)) == null) {
+            if (this.mBaseName.exists()) {
+                if (!this.mBackupName.exists()) {
+                    if (!this.mBaseName.renameTo(this.mBackupName)) {
+                        Log.w("AtomicFile", "Couldn't rename file " + this.mBaseName + " to backup file " + this.mBackupName);
+                    }
+                } else {
+                    this.mBaseName.delete();
                 }
             }
-            throw new IOException("Couldn't create directory " + this.mBaseName);
+            try {
+                return new FileOutputStream(this.mBaseName);
+            } catch (FileNotFoundException unused) {
+                if (this.mBaseName.getParentFile().mkdirs()) {
+                    try {
+                        return new FileOutputStream(this.mBaseName);
+                    } catch (FileNotFoundException unused2) {
+                        throw new IOException("Couldn't create " + this.mBaseName);
+                    }
+                }
+                throw new IOException("Couldn't create directory " + this.mBaseName);
+            }
         }
+        return (FileOutputStream) invokeV.objValue;
     }
 }
