+++ com/baidu/android/imsdk/retrieve/util/FileUtils.java
@@ -1,28 +1,55 @@
 package com.baidu.android.imsdk.retrieve.util;
 
+import com.baidu.titan.sdk.runtime.FieldHolder;
+import com.baidu.titan.sdk.runtime.InitContext;
+import com.baidu.titan.sdk.runtime.InterceptResult;
+import com.baidu.titan.sdk.runtime.Interceptable;
+import com.baidu.titan.sdk.runtime.TitanRuntime;
 import java.io.File;
 /* loaded from: classes.dex */
 public class FileUtils {
-    public static boolean deleteFile(File file) {
-        if (file == null) {
-            return false;
-        }
-        boolean z = true;
-        if (file.exists()) {
-            if (file.isFile()) {
-                return file.delete();
+    public static /* synthetic */ Interceptable $ic;
+    public transient /* synthetic */ FieldHolder $fh;
+
+    public FileUtils() {
+        Interceptable interceptable = $ic;
+        if (interceptable != null) {
+            InitContext newInitContext = TitanRuntime.newInitContext();
+            interceptable.invokeUnInit(65536, newInitContext);
+            int i2 = newInitContext.flag;
+            if ((i2 & 1) != 0) {
+                int i3 = i2 & 2;
+                newInitContext.thisArg = this;
+                interceptable.invokeInitBody(65536, newInitContext);
             }
-            if (file.isDirectory()) {
-                File[] listFiles = file.listFiles();
-                if (listFiles != null) {
-                    for (File file2 : listFiles) {
-                        z &= deleteFile(file2);
-                    }
+        }
+    }
+
+    public static boolean deleteFile(File file) {
+        InterceptResult invokeL;
+        Interceptable interceptable = $ic;
+        if (interceptable == null || (invokeL = interceptable.invokeL(65537, null, file)) == null) {
+            if (file == null) {
+                return false;
+            }
+            boolean z = true;
+            if (file.exists()) {
+                if (file.isFile()) {
+                    return file.delete();
                 }
-                return z & file.delete();
+                if (file.isDirectory()) {
+                    File[] listFiles = file.listFiles();
+                    if (listFiles != null) {
+                        for (File file2 : listFiles) {
+                            z &= deleteFile(file2);
+                        }
+                    }
+                    return z & file.delete();
+                }
+                return true;
             }
             return true;
         }
-        return true;
+        return invokeL.booleanValue;
     }
 }
